Refactored HealthComponent & Intergrated in player
- Fixed issue where HealthComponent appeared as a blank MonoScript in the Editor.
- Updated Character class to accept health initialization via SO_HealthConfig.

TL;DR -> All characters will have their own health configurations now & split up Component and Config scripts.
