+++ Assets/Script/HealthComponent.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 259307a5145c68448bf9e794bbe0dad2
