+++ Assets/Script/Scriptable Objects.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2f5c3850633167d4ab9ed2af59cb007c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
