+++ Assets/Script/Scriptable Objects/SO_DefaultHealthComponent.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b5fa22ad2408cc142ad6f341d2d2edc4
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
