+++ Assets/Script/Scriptable Objects/SO_WarriorHealth.asset
@@ -1,0 +1,18 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 9323f434171e2df4489668db18fef0d8, type: 3}
+  m_Name: SO_WarriorHealth
+  m_EditorClassIdentifier: 
+  maxHealth: 150
+  canRegenerate: 1
+  regenRate: 1
+  regenAmount: 1
