+++ Assets/Script/Scriptable Objects/SO_WarriorHealth.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 316a0f387aa42114fb5a48e82fc8596c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
