+++ Assets/prefab/CharacterObject.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ad88f7ef6b1d55c42a41f750bd9ee632
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
